==> code/3-analysis.Rmd <==
---
title: 'Agricultural projects: Project Appraisal Documents Descriptives'
output: pdf_document
---
  
```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = FALSE)
knitr::opts_chunk$set(message = FALSE)
```

```{r, include = FALSE}

library(tidyverse)
library(expss)
library(eulerr)
library(RColorBrewer)

github <- "C:/Users/wb501238/Documents/GitHub/operations-portal-scrape"
path_dt <- file.path(github,"data", "processed")

yearly <- readRDS(file.path(path_dt,"ag-projects-year-summary.RDS"))

projects <- 
  readRDS(file.path(path_dt, 
                    "ag-projects-constructed.RDS")) %>%
  filter(approvalfy > 1985) %>%
  filter(approvalfy <= 2020)


projects_region <- readRDS(file.path(path_dt, "ag-projects-region-summary.RDS"))

pads <-
  projects %>%
  filter(pad == TRUE)

components <- c("demo", "ext", "contact", "ffs")
themes     <- c("gender", "com", "climate")

```

# Dataset description
- We scraped the operations portal and WB docs to find the Project Appraisal Documents of World Bank project in the agriculture practice
- The operations portal lists a total of `r nrow(projects)` agriculture projects approved between `r min(projects$approvalfy)` and `r max(projects$approvalfy)` 
- Of those, `r nrow(pads)` projects included their PADs in the documentation on either the operations portal or WB docs

```{r}

data <-
  projects %>%
  group_by(approvalfy) %>%
  summarise(total = n(),
            pad = sum(pad, na.rm = TRUE)) %>%
  pivot_longer(cols = c("total", "pad"),
               names_to = "type") %>%
  mutate(type = type %>%
                str_replace_all("total", "Listed on operations portal") %>%
                str_replace_all("pad", "PAD available"))

ggplot(data,
       aes(x = approvalfy,
           y = value,
           color = type)) +
  geom_line(size = 1) + 
  geom_point(shape = 21,
             size = 2,
             fill = "white") +
  scale_color_manual(values = brewer.pal(7, "Dark2")[6:7]) +
  theme(legend.title = element_blank(),
        axis.ticks.x = element_blank(),
        legend.position = "top") +
  ylab("Number of projects") +
  xlab("Approval year") +
  scale_x_continuous(breaks = seq(1985, 2020, by = 5),
                     minor_breaks = seq(1985, 2020, by = 1)) 
```

# How common are mentions to components?

## Over time
```{r fig.height = 4}
 year <- 
  pads %>%
    group_by(approvalfy) %>%
    summarise_at(c(components, "pad"),
                 sum, na.rm = TRUE) %>%
    pivot_longer(cols = components,
                 names_to = "component") %>%
  mutate(pct = (value/pad) * 100,
           component = component %>% 
                       str_replace_all("demo", "Demonstration plot") %>%
                       str_replace_all("ext", "Agricultural extension") %>%
                       str_replace_all("contact", "Contact farmer") %>%
                       str_replace_all("ffs", "Farmer field school")
    )

 
  ggplot(data = year,
         aes(x = approvalfy,
             y = value,
             color = component,
             label = round(pct, digits = 1))) +
  geom_line(size = 1) + 
  geom_point(shape = 21,
             size = 2,
             fill = "white") +
  scale_color_brewer(palette = "Dark2") +
  theme(legend.title = element_blank(),
        axis.ticks.x = element_blank(),
        legend.position = "top") +
  ylab("Number of projects that\nmention component in PAD") +
  xlab("Approval year") +
  scale_x_continuous(breaks = seq(1990, 2020, by = 5),
                     minor_breaks = seq(1990, 2020, by = 1)) 
  
```

## By cross-cutting themes
```{r fig.height = 3}
 gender <- 
  pads %>%
    filter(gender == TRUE) %>%
    summarise_at(c(components, "pad"),
                 sum, na.rm = TRUE) %>%
    pivot_longer(cols = components,
                 names_to = "component") %>%
  mutate(theme = "Gender")

 com <- 
  pads %>%
    filter(com == TRUE) %>%
    summarise_at(c(components, "pad"),
                 sum, na.rm = TRUE) %>%
    pivot_longer(cols = components,
                 names_to = "component") %>%
   mutate(theme = "Commercialization")
 
 climate <- 
  pads %>%
    filter(climate == TRUE) %>%
    summarise_at(c(components, "pad"),
                 sum, na.rm = TRUE) %>%
    pivot_longer(cols = components,
                 names_to = "component") %>%
   mutate(theme = "Climate")


 themes <-
  bind_rows(gender, com, climate) %>%
  mutate(pct = (value/pad) * 100,
           component = component %>% 
                       str_replace_all("demo", "Demonstration plot") %>%
                       str_replace_all("ext", "Agricultural extension") %>%
                       str_replace_all("contact", "Contact farmer") %>%
                       str_replace_all("ffs", "Farmer field school")
    )

 
  ggplot(data = themes,
         aes(x = component,
             y = pct,
             color = component,
             label = round(pct, digits = 1),
             fill = component)) +
  geom_bar(stat = "identity") +
  geom_text(vjust = -0.3, 
            size = 2.5,
            color = "black")+
  scale_color_brewer(palette = "Dark2") +
  scale_fill_brewer(palette = "Dark2") +
  facet_wrap(~ theme) +
  theme(legend.title = element_blank(),
        axis.ticks.x = element_blank(),
        axis.text.x = element_blank(),
        legend.position = "top") +
  ylab("Percent of projects that\nmention component in PAD") +
  xlab(element_blank()) 
  
```

## Across regions
```{r fig.height = 5}

 region <- 
  pads %>%
    filter(region != "Other") %>%
    group_by(region) %>%
    summarise_at(c(components, "pad"),
                 sum, na.rm = TRUE) %>%
    pivot_longer(cols = components,
                 names_to = "component") %>%
    mutate(pct = (value/pad) * 100,
           component = component %>% 
                       str_replace_all("demo", "Demonstration plot") %>%
                       str_replace_all("ext", "Agricultural extension") %>%
                       str_replace_all("contact", "Contact farmer") %>%
                       str_replace_all("ffs", "Farmer field school")
    )

  ggplot(data = region,
         aes(x = component,
             y = pct,
             color = component,
             label = round(pct, digits = 1),
             fill = component)) +
  geom_bar(stat = "identity") +
  geom_text(vjust = -0.3, 
            size = 2.5,
            color = "black")+
  scale_color_brewer(palette = "Dark2") +
  scale_fill_brewer(palette = "Dark2") +
  facet_wrap(~ region) +
  theme(legend.title = element_blank(),
        axis.ticks.x = element_blank(),
        axis.text.x = element_blank(),
        legend.position = "top") +
  ylab("Percent of projects that mention component in PAD") +
  xlab(element_blank()) + 
  ylim(0, 80)
  
```


# How often are farmer field school projects implemented in partnership with FAO? 

## Common mentions
```{r fig.height = 3}

fao_projects <-
  projects %>%
  filter(fao == TRUE) %>%
  select(id) %>%
  unlist %>%
  unname

ffs_projects <-
  projects %>%
  filter(ffs == TRUE) %>%
  select(id) %>%
  unlist %>%
  unname

fao_projects <-
  projects %>%
  filter(fao == TRUE) %>%
  select(id) %>%
  unlist %>%
  unname

ffs_projects <-
  projects %>%
  filter(ffs == TRUE) %>%
  select(id) %>%
  unlist %>%
  unname

euler <- list(`Farmer field school` = ffs_projects,
              `FAO` = fao_projects)

colors <- brewer.pal(6, "Dark2")[4:6]

plot(euler(euler),
     fills = list(fill = colors, 
                  alpha = .5),
     edges = FALSE,
     quantities = list(fontefamily = "mono"),
     counts = TRUE)

```


## Trends
```{r fig.height = 3.5}
fao <-
  yearly %>%
  filter(expression %in% c("Farmer field school", "FAO")) %>%
  filter(pads > 0)

  ggplot(data = fao,
         aes(x = year,
             y = count,
             color = expression)) +
  geom_line(size = 1) + 
  geom_point(shape = 21,
             size = 2,
             fill = "white") +
    scale_color_manual(values = c(brewer.pal(5,"Dark2")[5], 
                     brewer.pal(5,"Dark2")[4])) +
    theme(legend.title = element_blank(),
        axis.ticks.x = element_blank(),
        legend.position = "top") +
  ylab("Number of projects that\nmention component in PAD") +
  xlab("Approval year") +
  scale_x_continuous(breaks = seq(1990, 2020, by = 5),
                     minor_breaks = seq(1990, 2020, by = 1)) 
```

# How are components presence related to project net commitment?

## Total net commitment
```{r fig.height = 3.5}
value_year <- 
  pads %>%
  mutate(netcommitment = as.numeric(netcommitment)) %>%
  group_by(approvalfy) %>%
  summarise(median = median(netcommitment),
              total = sum(netcommitment),
              mean = mean(netcommitment)) %>%
  mutate(component = "All projects")

value <- 
  pads %>%
    select(approvalfy, netcommitment, demo, ext, contact, ffs) %>%
    pivot_longer(cols = components,
                 names_to = "component") %>%
    filter(value == TRUE) %>%
    group_by(approvalfy, component) %>%
  mutate(netcommitment = as.numeric(netcommitment)) %>%
    summarise(median = median(netcommitment),
              total = sum(netcommitment),
              mean = mean(netcommitment)) %>%
  mutate(component = component %>% 
                       str_replace_all("demo", "Demonstration plot") %>%
                       str_replace_all("ext", "Agricultural extension") %>%
                       str_replace_all("contact", "Contact farmer") %>%
                       str_replace_all("ffs", "Farmer field school")) %>%
  bind_rows(value_year)

 
  ggplot(data = value %>%
                filter(component != "All projects"),
         aes(x = approvalfy,
             y = total,
             color = component)) +
  geom_line(size = 1) + 
  geom_point(shape = 21,
             size = 2,
             fill = "white") +
  scale_color_brewer(palette = "Dark2") +
  geom_line(data = value %>%
                   filter(component == "All projects"),
            size = 1,
            color = brewer.pal(6, "Dark2")[6],
            linetype = "dashed") + 
  annotate("text", 
           x = 2019, 
           y = 2000, 
           label= "Overall",
           size = 2.5) +
  theme(legend.title = element_blank(),
        axis.ticks.x = element_blank(),
        legend.position = "top") +
  ylab("USD Million, current") +
  xlab("Approval year") +
  scale_x_continuous(breaks = seq(1990, 2020, by = 5),
                     minor_breaks = seq(1990, 2020, by = 1))
```

## Median net commitment
```{r fig.height = 3.5}
  ggplot(data = value %>%
                filter(component != "All projects"),
         aes(x = approvalfy,
             y = median,
             color = component)) +
  geom_line(size = 1) + 
  geom_point(shape = 21,
             size = 2,
             fill = "white") +
  scale_color_brewer(palette = "Dark2") +
  geom_line(data = value %>%
                   filter(component == "All projects"),
            size = 1,
            color = brewer.pal(6, "Dark2")[6],
            linetype = "dashed") + 
annotate("text", 
         x = 1997, 
         y = 100, 
         label= "Overall", 
         size = 2.5) +
theme(legend.title = element_blank(),
        axis.ticks.x = element_blank(),
        legend.position = "top") +
  ylab("USD Million, current") +
  xlab("Approval year") +
  scale_x_continuous(breaks = seq(1990, 2020, by = 5),
                     minor_breaks = seq(1990, 2020, by = 1))
  
```
